refactor(adr-061): split source_graph construction/comparison (Phase 5 item 2)
Finishes ADR-061 Phase 5 item 2's remaining two-thirds: buildsource/
source_graph.py (1352 lines) becomes a pure re-export facade (140 lines).

- Construction (build_source_graph + private folding helpers) splits into
  two new flat modules, purely to stay under the 800-line new-file cap:
  source_graph_build.py (Phase 2: BuildEvidence folding) and
  source_graph_build_source_abi.py (Phases 3-4 + the ADR-038 C.9
  source_edges fold). Both classified extract in architecture/modules.yaml.
- Comparison (diff_source_graph, localize_symbol) moves into
  source_graph_compare.py, classified compare.
- A third slice the original two-way split didn't name: the shared
  node/edge-classification predicates (is_public_dependency_node,
  is_internal_dependency_node, looks_like_system_name, etc.) used well
  beyond either half (crosscheck.py, graph_reconcile.py, internal_leak.py,
  impact/*, surface.py) move into source_graph_query.py, left unclassified
  in modules.yaml -- the same state several of its own callers are in.

Re-measured the build_evidence.py/comdat_groups.py blocker this ADR
recorded rather than assuming it still applied: it only ever blocked
physically relocating graph_facts.py/entity_resolver.py into
abicheck/model/ (a model -> extract cycle risk). It never applied to
construction/comparison staying flat in buildsource/ --
check_architecture.py's dependency-direction checks only fire for modules
physically migrated under a layer's own directory. Confirmed via
check_architecture.py: 0 errors before and after.

source_graph_findings.py's own imports were switched from the facade to
the real new homes (a genuine reverse-facade-import bug fix, not just a
move: its target names left the facade's module body entirely). Every
other pre-existing internal caller (~15 flat buildsource/ modules) keeps
importing through the facade, matching the 'values' slice's own precedent
of not rewriting all 77 existing callers in one PR.

Tests split alongside the implementation (D10): tests/test_source_graph.py
(1996 lines) -> construction tests to test_source_graph_build.py (885
lines), comparison tests to test_source_graph_compare.py (143 lines),
everything else (findings/schema/pack/CLI) stays (1175 lines). All 105
tests across the three files pass unchanged.

architecture/debt.yaml's source_graph.py and test_source_graph.py
no-growth entries removed (rationale no longer applies at 140/1175 lines).

Verification: check_architecture.py 0 errors, check_ai_readiness.py no new
errors (3 pre-existing ADR-verification errors confirmed present on base
branch), ruff check clean, mypy clean, adr_status_sync.py clean. Targeted
tests (source_graph/build_evidence/comdat): 239 passed. Broader scoped run
(138 test files covering source_graph and every downstream consumer --
crosscheck, graph_reconcile, internal_leak, impact/*, call_graph,
template_graph, type_graph, and more): 6606 passed, 62 skipped, 0 failed.
