+++ abicheck/buildsource/source_graph.py
@@ -12,30 +12,54 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Source / implementation graph summary (ADR-031 L5).
+"""Back-compat facade: source-graph values/construction/comparison split.
 
-abicheck's own normalized, ABI/API-relevant graph. Stored compactly as
-``graph/source_graph_summary.json`` inside an evidence pack (ADR-028 D8): the
-primary snapshot only ever keeps a coverage row + reference, never the full
-graph (ADR-031 D1, D7).
+ADR-061 Phase 5 item 2's construction/comparison split (following the same
+"values" move this module's own docstring already tracked): graph values
+(``SourceGraphSummary``/``GraphSummaryDiff``, node-id constructors, schema
+vocabulary) live in ``abicheck.model.source_graph``; ``GraphNode``/
+``GraphEdge`` + the ADR-046 fact-merge machinery live in
+``abicheck.model.graph_facts``. Construction (:func:`build_source_graph` and
+its private helpers, folding an ADR-029 ``BuildEvidence`` + optional ADR-030
+``SourceAbiSurface`` into a graph) is split across ``source_graph_build.py``
+(Phase 2) and ``source_graph_build_source_abi.py`` (Phases 3-4 + the ADR-038
+C.9 ``source_edges`` fold) — two flat modules rather than one, purely to stay
+under the new-file line-count cap. Comparison (:func:`diff_source_graph`,
+:func:`localize_symbol`) lives in ``source_graph_compare.py``. The shared
+node/edge-classification predicates neither half owns exclusively (used by
+``crosscheck.py``, ``graph_reconcile.py``, ``internal_leak.py``,
+``impact/*``, ``surface.py``, ``post_processing_reachability.py``, and
+others) live in ``source_graph_query.py``.
 
-This module implements the MVP scope of the ADR:
-
-- **Phase 1** — the node/edge schema, the compact ``SourceGraphSummary``
-  container, content addressing, and round-trip (de)serialization.
-- **Phase 2** — :func:`build_source_graph`, which folds an ADR-029
-  :class:`~abicheck.buildsource.build_evidence.BuildEvidence` into a
-  target/source/header/compile-unit/build-option graph.
-- A structural :func:`diff_source_graph` (Phase 5 seed) that powers the
-  ``graph compare`` command for explanation and triage.
-
-Every edge carries provenance and a confidence label (ADR-031 D2, D9): a graph fact must always say *how* it was derived so a reader never mistakes graph absence for safety.
-Deeper layers — public-reachability / type / include / call graphs (Phases 3-4, 6) and external backends like Kythe/CodeQL (Phase 7) — extend this same schema; per ADR-031 D6 graph diffs *explain and prioritize* and must never, on their own, silently decide or suppress an artifact-proven ABI break (ADR-028 D3).
+This module re-exports every moved name (``X as X``, the same convention its
+own pre-existing ``graph_facts``/model re-export blocks already used) so
+every existing ``from .source_graph import ...``/``from
+abicheck.buildsource.source_graph import ...`` call site — inside this
+package and out — keeps resolving unchanged.
 """
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any
+from typing import Any
+
+# GraphNode/GraphEdge live in graph_facts.py now (ADR-046 D1/D2/D3 schema
+# additions pushed this module to its AI-readiness line-count cap) and are
+# re-exported for backward compatibility (many modules do ``from
+# .source_graph import GraphNode``/``CONF_HIGH`` etc.) — the ``as``-aliases
+# make the re-export explicit for mypy's strict ``--no-implicit-reexport``.
+from ..model.graph_facts import (
+    CONF_HIGH as CONF_HIGH,
+    CONF_REDUCED as CONF_REDUCED,
+    CONF_UNKNOWN as CONF_UNKNOWN,
+    FactConflict as FactConflict,
+    GraphEdge as GraphEdge,
+    GraphFact as GraphFact,
+    GraphNode as GraphNode,
+    _decl_node_id as _decl_node_id,
+    _normalize_graph_identity as _normalize_graph_identity,
+    _type_node_id as _type_node_id,
+    register_fact as register_fact,
+)
 
 # SourceGraphSummary/GraphSummaryDiff, the node-id constructors, and the
 # schema vocabulary (NODE_KINDS/EDGE_KINDS/...) live in
@@ -64,1273 +88,37 @@ from ..model.source_graph import (
     _vtable_node_id as _vtable_node_id,
     function_decl_identity as function_decl_identity,
 )
-from .build_evidence import BuildEvidence, Confidence
 
-# GraphNode/GraphEdge live in graph_facts.py now (ADR-046 D1/D2/D3 schema
-# additions pushed this module to its AI-readiness line-count cap) and are
-# re-exported for backward compatibility (many modules do ``from
-# .source_graph import GraphNode``/``CONF_HIGH`` etc.) — the ``as``-aliases
-# make the re-export explicit for mypy's strict ``--no-implicit-reexport``.
-from .graph_facts import (
-    CONF_HIGH as CONF_HIGH,
-    CONF_REDUCED as CONF_REDUCED,
-    CONF_UNKNOWN as CONF_UNKNOWN,
-    FactConflict as FactConflict,
-    GraphEdge as GraphEdge,
-    GraphFact as GraphFact,
-    GraphNode as GraphNode,
-    _decl_node_id as _decl_node_id,
-    _normalize_graph_identity as _normalize_graph_identity,
-    _type_node_id as _type_node_id,
-    register_fact,
+# Construction (ADR-061 Phase 5 item 2's construction slice).
+from .source_graph_build import (
+    build_source_graph as build_source_graph,
+    project_source_files as project_source_files,
+)
+from .source_graph_build_source_abi import (
+    _file_in_project as _file_in_project,
+    fold_source_edges as fold_source_edges,
+    mark_source_edges_extractor_coverage as mark_source_edges_extractor_coverage,
 )
 
-if TYPE_CHECKING:
-    from .source_abi import SourceAbiSurface, SourceEntity
-
-
-def _conf_from_build(conf: Confidence) -> str:
-    """Map an ADR-029 build-evidence confidence onto a graph confidence label."""
-    if conf == Confidence.HIGH:
-        return CONF_HIGH
-    if conf == Confidence.REDUCED:
-        return CONF_REDUCED
-    return CONF_UNKNOWN
-
-
-#: Suffixes identifying a static-library archive among a LinkUnit's inputs
-#: (ADR-041 P1 #2). Lowercase only — compared case-insensitively below (Codex
-#: review): Windows evidence can spell this uppercase (``FOO.LIB``), hidden from ``archive_graph.py`` otherwise, same as ``adapters/make.py``.
-_STATIC_LIBRARY_SUFFIXES = (".a", ".lib")
-
-
-#: Graph node kinds a type entity (as opposed to a function/variable
-#: ``source_decl``) can carry. Mirrors ``crosscheck._DECL_NODE_KINDS`` minus ``source_decl``.
-_TYPE_ENTITY_KINDS: frozenset[str] = frozenset({"record_type", "enum_type", "typedef"})
-
-#: Graph node kinds that carry a declaration/type visibility we can classify as
-#: public or internal. Shared with ``crosscheck.py``'s intra-version
-#: ``public_to_internal_dependency`` check (ADR-041 P0 slice 2, fourth Codex
-#: review) so the two never classify a node differently.
-DECL_NODE_KINDS: frozenset[str] = frozenset({"source_decl"}) | _TYPE_ENTITY_KINDS
-
-#: Node visibilities that put an entity *on* the public source surface. Mirrors
-#: ``source_link._is_public`` (which the L5 graph's ``visibility`` attr is
-#: derived from): ``generated`` means a generated header **under the public
-#: roots** — a public, consumer-visible entity — so it is NOT an internal dependency.
-PUBLIC_VISIBILITIES: frozenset[str] = frozenset({"public_header", "generated"})
-
-#: Node visibilities that make an entity *internal* (not public surface): a
-#: project-private header or an implementation ("source") file. System headers
-#: are third-party (excluded), and ``generated`` is public (above).
-INTERNAL_VISIBILITIES: frozenset[str] = frozenset({"private_header", "source"})
-
-#: Visibilities that carry no provenance. The built-in call/type-graph
-#: extractors create dependency-target nodes with **no** ``visibility`` attr
-#: when the target isn't part of the linked L4 surface. Such a node is
-#: internal *only when the project also declares it* (``decl_to_file``) or the
-#: extractor marked it ``defined_in_project`` — caller/reference presence
-#: alone is unsound (a third-party header-inline symbol whose body is reached
-#: also appears as a dependency target), so a bare node with no project
-#: provenance is treated as a third-party/system target and not flagged.
-UNANNOTATED_VISIBILITIES: frozenset[str] = frozenset({"", "unknown"})
-
-#: Mangled-name prefixes / substrings that mark a standard-library or
-#: compiler-internal decl. The call/type graphs resolve targets into ``std::``/
-#: ``__gnu_cxx``/cxxabi helpers, which carry no visibility either; without this
-#: an unannotated stdlib target would be mis-read as a project-internal
-#: dependency and a public API merely using ``std::`` would light up. Mirrors
-#: the stdlib/compiler filtering the dumper already applies to exported
-#: symbols.
-_SYSTEM_NAME_PREFIXES = (
-    "_ZSt",
-    "_ZNSt",
-    "_ZNKSt",
-    "_ZNSa",
-    "_ZN9__gnu_cxx",
-    "_ZNK9__gnu_cxx",
-    "_ZN6__cxxabiv",
-    "_Znw",
-    "_Zna",
-    "_Zdl",
-    "_Zda",
-    "__",
-)
-_SYSTEM_NAME_SUBSTRINGS = ("std::", "__gnu_cxx::", "__cxxabiv")
-
-
-def looks_like_system_name(name: str) -> bool:
-    """Whether *name* is a standard-library / compiler-internal decl spelling."""
-    if name.startswith(_SYSTEM_NAME_PREFIXES):
-        return True
-    return any(sub in name for sub in _SYSTEM_NAME_SUBSTRINGS)
-
-
-def decl_declaring_files(graph: SourceGraphSummary) -> dict[str, str]:
-    """Map each decl/type id to its declaring file via ``SOURCE_DECLARES`` edges."""
-    node_by_id = {n.id: n for n in graph.nodes}
-    decl_to_file: dict[str, str] = {}
-    for e in graph.edges:
-        if e.kind != "SOURCE_DECLARES":
-            continue
-        header = node_by_id.get(e.src)
-        if header is not None and header.label:
-            decl_to_file.setdefault(e.dst, header.label)
-    return decl_to_file
-
-
-def is_public_dependency_node(
-    node_id: str, node_by_id: dict[str, GraphNode], exported_decls: set[str]
-) -> bool:
-    """Whether *node_id* is public: exported-symbol-mapped or public-header visible.
-
-    Shared with ``crosscheck.py``'s ``_is_public_decl`` (ADR-041 P0 slice 2).
-    Deliberately does not consider whether the node's own body is compiled
-    into consumer code (see :func:`is_consumer_compiled_public_entry`) — an
-    exported-or-header-visible declaration is exactly the "public API
-    surface" question ``crosscheck.py``'s advisory
-    ``public_to_internal_dependency`` check (RISK-only, never gates
-    suppression) wants to ask, regardless of where the declaration's body
-    lives.
-    """
-    if node_id in exported_decls:
-        return True
-    node = node_by_id.get(node_id)
-    if node is None or node.kind not in DECL_NODE_KINDS:
-        return False
-    return str(node.attrs.get("visibility", "")) in PUBLIC_VISIBILITIES
-
-
-def is_consumer_compiled_public_entry(
-    node_id: str, node_by_id: dict[str, GraphNode], exported_decls: set[str]
-) -> bool:
-    """Whether *node_id* is a public entry whose own body is compiled into
-    consumer binaries — the correct "entry" set for a *call-graph*
-    reachability walk (Codex review, fresh evidence).
-
-    :func:`is_public_dependency_node` alone over-reaches here: an ordinary,
-    out-of-line exported function (e.g. ``api()`` defined in a ``.cpp``
-    file) is public, but its *body* — and therefore its own internal calls,
-    e.g. to ``ns::detail::helper()`` — is compiled into the **library's**
-    binary only, never into any consumer's. A consumer links against
-    ``api()``'s exported symbol alone; it never sees, references, or
-    embeds ``helper()``. So walking the call graph from *every* exported
-    function (as :func:`is_public_dependency_node` does) treats an ordinary
-    internal implementation-detail call as if it were public-reachable,
-    which either manufactures a spurious "still reachable" narrative on a
-    genuinely safe-to-suppress internal change, or (via
-    ``post_processing.MarkReachability``) blocks a broad internal-namespace
-    suppression rule from ever applying to the common case — most
-    functions in most libraries are ordinary, out-of-line, non-template.
-
-    The real criterion is whether the entry's own body is emitted into
-    every including translation unit — true for inline functions/methods
-    and templates, false for an ordinary out-of-line definition — captured
-    by ``GraphNode.attrs["consumer_compiled_body"]``
-    (:func:`build_source_graph`). A node without that attr at all defaults
-    permissively to ``True`` — matching the header-graph/type-node/generic
-    case, where no signal either way is available — **except** a node whose
-    ``provenance`` is one of :data:`_NO_CONSUMER_COMPILED_SIGNAL_PROVENANCES`
-    (Codex review, fresh evidence): ``augment_graph_with_calls``
-    (``call_graph.py``) stamps its own fallback tag on a node it creates for
-    a caller/callee identity with no other declaration node backing it — a
-    real, build-integrated project function reached only through the call
-    graph itself, whose out-of-line body is not necessarily
-    consumer-compiled. An inline public ``wrap()`` calling an ordinary
-    out-of-line project function ``helper_a()`` (this fallback shape) which
-    itself calls an internal ``ns::detail::helper()`` must stop expanding
-    *at* ``helper_a()`` — treating "no signal" as "safe" for this one node
-    shape would silently reintroduce the exact over-reach this predicate
-    exists to reject. ``graph_backends.py``'s Kythe/CodeQL ingestion
-    (``ingest_kythe_entries``/``ingest_codeql_call_results``) creates the
-    identical shape for an external-indexer edge: a bare ``source_decl``
-    node stamped with provenance ``"kythe"``/``"codeql"`` and no
-    ``consumer_compiled_body`` attr at all, since neither export format says
-    whether the referenced declaration's body is inline/template — an
-    imported Kythe/CodeQL call chain through an ordinary out-of-line
-    intermediate helper must stop there too, for the same reason.
-    """
-    if not is_public_dependency_node(node_id, node_by_id, exported_decls):
-        return False
-    return is_consumer_compiled_node(node_id, node_by_id)
-
-
-#: ``provenance`` tag ``augment_graph_with_calls`` (``call_graph.py``) stamps on a fallback
-#: node it creates for a caller/callee identity with no other declaration node backing it —
-#: the one node shape known to lack a ``consumer_compiled_body`` attr while still representing
-#: a genuine, build-integrated (out-of-line, not-necessarily-consumer-compiled) project
-#: declaration, as opposed to "no signal available" (header-graph/type nodes, synthetic test
-#: fixtures, …) which stays permissive by default. Mirrored as a literal string rather than
-#: imported from ``call_graph.py`` to avoid coupling this module to that one's own constant.
-_CALL_GRAPH_FALLBACK_PROVENANCE = "call_graph"
-
-#: Same shape as :data:`_CALL_GRAPH_FALLBACK_PROVENANCE`, for external
-#: indexer backends (Codex review, fresh evidence): ``graph_backends.py``'s
-#: ``ingest_kythe_entries``/``ingest_codeql_call_results``/
-#: ``ingest_codeql_extends_results`` stamp exactly these two provenance
-#: strings on a bare ``source_decl``/``record_type`` node with no
-#: ``consumer_compiled_body`` attr — Kythe entries and CodeQL query results
-#: carry cross-reference edges only, never whether the referenced
-#: declaration's body is inline/template, so an attr-less node reached only
-#: through one of these backends is exactly as unproven as the call-graph
-#: fallback shape and must not be treated as a safe stopping point by
-#: default either.
-_NO_CONSUMER_COMPILED_SIGNAL_PROVENANCES = frozenset(
-    {
-        _CALL_GRAPH_FALLBACK_PROVENANCE,
-        "kythe",
-        "codeql",
-    }
+# Comparison (ADR-061 Phase 5 item 2's comparison slice).
+from .source_graph_compare import (
+    diff_source_graph as diff_source_graph,
+    localize_symbol as localize_symbol,
 )
 
-
-def is_consumer_compiled_node(node_id: str, node_by_id: dict[str, GraphNode]) -> bool:
-    """Whether *node_id*'s own body is compiled into consumer code, independent
-    of whether it also qualifies as a *public* entry (see
-    :func:`is_consumer_compiled_public_entry` for that combined check) — the
-    predicate a call-graph *traversal* needs at every intermediate node, not
-    just at the entries it starts from (Codex review, fresh evidence: see
-    :func:`is_consumer_compiled_public_entry`'s docstring for the fallback-node
-    shapes this conservative exception protects against).
-    """
-    node = node_by_id.get(node_id)
-    if node is None:
-        return True
-    if "consumer_compiled_body" in node.attrs:
-        return bool(node.attrs["consumer_compiled_body"])
-    return node.provenance not in _NO_CONSUMER_COMPILED_SIGNAL_PROVENANCES
-
-
-def is_internal_dependency_node(
-    node_id: str,
-    node_by_id: dict[str, GraphNode],
-    exported_decls: set[str],
-    decl_to_file: dict[str, str],
-) -> bool:
-    """Whether *node_id* is a project-internal decl/type consumers cannot see.
-
-    "Not declared by a public header" alone is not internal — a third-party or
-    standard-library type used as a field/parameter type is *also* not
-    declared by any project header, and must not be conflated with a genuinely
-    private project entity (ADR-041 P0 slice 2, fourth Codex review). Requires
-    positive evidence instead: an explicit ``private_header``/``source``
-    visibility, or — for an unannotated node — project-file provenance
-    (``decl_to_file``/``defined_in_project``) plus a non-system-looking name.
-    Shared with ``crosscheck.py``'s ``_is_internal_decl`` (same algorithm, same
-    source of truth) so the intra-version and inter-version checks classify a
-    node identically.
-    """
-    node = node_by_id.get(node_id)
-    if node is None or node.kind not in DECL_NODE_KINDS:
-        return False
-    if node_id in exported_decls:
-        return False
-    vis = str(node.attrs.get("visibility", ""))
-    if vis in INTERNAL_VISIBILITIES:
-        return True
-    if vis in UNANNOTATED_VISIBILITIES:
-        has_provenance = node_id in decl_to_file or bool(
-            node.attrs.get("defined_in_project")
-        )
-        if not has_provenance:
-            return False
-        return not looks_like_system_name(node.label or "")
-    return False
-
-
-# ── Phase 2: build the graph from ADR-029 BuildEvidence ─────────────────────
-
-
-def _file_in_project(caller_file: str, project_files: frozenset[str]) -> bool:
-    """Whether *caller_file* is one of the project's own compile-unit sources.
-
-    Build-evidence sources are often repo-relative (``src/foo.cc``) while the
-    clang AST emits an absolute path (``/work/src/foo.cc``); match on a path
-    suffix either way (mirrors ``source_replay._path_matches``). A function whose
-    body is in one of these files is project-defined; one in a third-party/system
-    header (Boost/Abseil/libstdc++) is not.
-    """
-    if not caller_file:
-        return False
-    c = caller_file.replace("\\", "/").lstrip("./")
-    for pf in project_files:
-        n = pf.replace("\\", "/").lstrip("./")
-        if c == n or c.endswith("/" + n) or n.endswith("/" + c):
-            return True
-    return False
-
-
-def project_source_files(build: BuildEvidence) -> frozenset[str]:
-    """Project-internal source files for ``defined_in_project`` provenance.
-
-    Compile-unit sources **plus the targets' private headers** — a function whose
-    body is in a project ``.cc`` *or* a project private header is internal
-    implementation. Public headers are deliberately excluded: an inline function
-    in a public header is consumer-visible public surface, so marking it
-    ``defined_in_project`` (→ internal) would false-positive
-    ``public_to_internal_dependency``. Third-party/system headers (Boost, libc++)
-    are never in either list, so they stay external (Codex review).
-    """
-    files: set[str] = {cu.source for cu in build.compile_units if cu.source}
-    for tgt in build.targets:
-        files.update(h for h in tgt.private_headers if h)
-    return frozenset(files)
-
-
-def build_source_graph(
-    build: BuildEvidence, source_abi: SourceAbiSurface | None = None
-) -> SourceGraphSummary:
-    """Fold ADR-029 build evidence (+ optional L4 source surface) into a graph.
-
-    **Phase 2** emits the build-level slice from *build*:
-
-    - ``target`` nodes, with ``TARGET_HAS_SOURCE`` / ``TARGET_HAS_PUBLIC_HEADER``
-      / ``TARGET_DEPENDS_ON`` edges;
-    - ``compile_unit`` nodes, with ``COMPILE_UNIT_BUILDS_SOURCE`` edges and
-      ``COMPILE_UNIT_USES_OPTION`` edges to the ABI-relevant flags they carry;
-    - ``source`` / ``header`` / ``generated_file`` nodes (a source listed in
-      ``build.generated_files`` is typed ``generated_file``).
-
-    **Phases 3-4** — when an ADR-030 ``source_abi`` surface is supplied — add the
-    public-reachability and source↔binary slices: ``source_decl`` / type / macro
-    nodes declared by public headers (``SOURCE_DECLARES``), their
-    ``SOURCE_DECL_MAPS_TO_SYMBOL`` / ``SOURCE_TYPE_MAPS_TO_DEBUG_TYPE`` mappings,
-    and ``BINARY_EXPORTS_SYMBOL`` edges from the owning target. Together they
-    yield the target → public-header → decl → exported-symbol closure that
-    reachability triage needs.
-
-    Deeper call edges and external backends (Phases 6-7) extend the same graph.
-    """
-    graph = SourceGraphSummary()
-    generated = set(build.generated_files)
-
-    def file_node(path: str, *, header: bool = False) -> str:
-        if not path:
-            return ""
-        if path in generated:
-            node_id = _source_node_id(path)
-            graph.add_node(
-                GraphNode(
-                    id=node_id,
-                    kind="generated_file",
-                    label=path,
-                    provenance="build_evidence",
-                    confidence=CONF_REDUCED,
-                    attrs={"generated": True},
-                )
-            )
-            return node_id
-        if header:
-            node_id = _header_node_id(path)
-            graph.add_node(
-                GraphNode(
-                    id=node_id,
-                    kind="header",
-                    label=path,
-                    provenance="build_evidence",
-                    confidence=CONF_HIGH,
-                )
-            )
-            return node_id
-        node_id = _source_node_id(path)
-        graph.add_node(
-            GraphNode(
-                id=node_id,
-                kind="source",
-                label=path,
-                provenance="build_evidence",
-                confidence=CONF_HIGH,
-            )
-        )
-        return node_id
-
-    known_targets = {t.id for t in build.targets}
-    for tgt in build.targets:
-        conf = _conf_from_build(tgt.confidence)
-        graph.add_node(
-            GraphNode(
-                id=tgt.id,
-                kind="target",
-                label=tgt.name or tgt.id,
-                provenance="build_evidence",
-                confidence=conf,
-                attrs={
-                    "kind": tgt.kind.value,
-                    "visibility": tgt.visibility,
-                    "build_system": tgt.build_system,
-                },
-            )
-        )
-        for src in tgt.source_files:
-            sid = file_node(src)
-            graph.add_edge(
-                GraphEdge(
-                    src=tgt.id,
-                    dst=sid,
-                    kind="TARGET_HAS_SOURCE",
-                    provenance="build_evidence",
-                    confidence=conf,
-                )
-            )
-        for hdr in tgt.public_headers:
-            hid = file_node(hdr, header=True)
-            graph.add_edge(
-                GraphEdge(
-                    src=tgt.id,
-                    dst=hid,
-                    kind="TARGET_HAS_PUBLIC_HEADER",
-                    provenance="build_evidence",
-                    confidence=conf,
-                )
-            )
-        for dep in tgt.dependencies:
-            # Reference an external dependency explicitly when it is not one of
-            # our own targets, so the graph distinguishes intra-project edges
-            # from third-party ones (informative for reachability triage).
-            if dep not in known_targets:
-                graph.add_node(
-                    GraphNode(
-                        id=dep,
-                        kind="external_dependency",
-                        label=dep,
-                        provenance="build_evidence",
-                        confidence=CONF_REDUCED,
-                    )
-                )
-            graph.add_edge(
-                GraphEdge(
-                    src=tgt.id,
-                    dst=dep,
-                    kind="TARGET_DEPENDS_ON",
-                    provenance="build_evidence",
-                    confidence=conf,
-                )
-            )
-
-    for cu in build.compile_units:
-        graph.add_node(
-            GraphNode(
-                id=cu.id,
-                kind="compile_unit",
-                label=cu.output or cu.source or cu.id,
-                provenance="build_evidence",
-                confidence=CONF_HIGH,
-                attrs={
-                    "language": cu.language,
-                    "standard": cu.standard,
-                    "target_id": cu.target_id,
-                },
-            )
-        )
-        if cu.source:
-            sid = file_node(cu.source)
-            graph.add_edge(
-                GraphEdge(
-                    src=cu.id,
-                    dst=sid,
-                    kind="COMPILE_UNIT_BUILDS_SOURCE",
-                    provenance="build_evidence",
-                    confidence=CONF_HIGH,
-                )
-            )
-        for flag in cu.abi_relevant_flags:
-            oid = _option_node_id(flag)
-            graph.add_node(
-                GraphNode(
-                    id=oid,
-                    kind="build_option",
-                    label=flag,
-                    provenance="build_evidence",
-                    confidence=CONF_HIGH,
-                    attrs={"abi_relevant": True},
-                )
-            )
-            graph.add_edge(
-                GraphEdge(
-                    src=cu.id,
-                    dst=oid,
-                    kind="COMPILE_UNIT_USES_OPTION",
-                    provenance="build_evidence",
-                    confidence=CONF_HIGH,
-                )
-            )
-
-    _fold_link_provenance(graph, build)
-
-    if source_abi is not None:
-        _augment_with_source_abi(graph, source_abi, project_source_files(build))
-        _link_options_to_symbols(graph)
-
-    return graph.finalize()
-
-
-def _link_options_to_symbols(graph: SourceGraphSummary) -> None:
-    """Add ``BUILD_OPTION_AFFECTS_SYMBOL`` edges (ADR-031 D2, build→symbol flow).
-
-    Connects each ABI-relevant build option to the exported symbols it can
-    affect, via the path *option ← compile_unit (target) → exported symbol*.
-    Only meaningful once the L4 surface has contributed ``BINARY_EXPORTS_SYMBOL``
-    edges, so it is a no-op for a build-only graph.
-    """
-    target_syms: dict[str, list[str]] = {}
-    for e in graph.edges:
-        if e.kind == "BINARY_EXPORTS_SYMBOL":
-            target_syms.setdefault(e.src, []).append(e.dst)
-    if not target_syms:
-        return
-    cu_target = {
-        n.id: str(n.attrs.get("target_id", ""))
-        for n in graph.nodes
-        if n.kind == "compile_unit"
-    }
-    for e in list(graph.edges):
-        if e.kind != "COMPILE_UNIT_USES_OPTION":
-            continue
-        target = cu_target.get(e.src, "")
-        for sym in target_syms.get(target, []):
-            graph.add_edge(
-                GraphEdge(
-                    src=e.dst,
-                    dst=sym,
-                    kind="BUILD_OPTION_AFFECTS_SYMBOL",
-                    provenance="build_evidence+source_abi",
-                    confidence=CONF_REDUCED,
-                )
-            )
-
-
-def _fold_link_provenance(graph: SourceGraphSummary, build: BuildEvidence) -> None:
-    """Fold object/link provenance from *build* into *graph* (ADR-041 P1 #2).
-
-    Lets a symbol change be attributed to "which object/archive member/link
-    step", not only "which target" — the gap the roadmap named:
-    ``TARGET_DEPENDENCY_ADDED``/``EXPORTED_SYMBOL_SOURCE_OWNER_CHANGED``
-    currently cannot explain an accidental export from a static archive, a
-    COMDAT/weak-symbol resolution change, or a transitive ``DT_NEEDED`` traced
-    to a specific object.
-
-    - Every ``compile_unit`` with a known ``output`` gets an ``object_file``
-      node and a ``COMPILE_UNIT_EMITS_OBJECT`` edge — "this TU produced this
-      object."
-    - Every :class:`~abicheck.buildsource.build_evidence.LinkUnit` becomes a
-      ``link_unit`` node (``NODE_KINDS`` reserved this kind since ADR-031 D2
-      but nothing populated it before this), linked to its owning ``target``
-      (``TARGET_HAS_LINK_UNIT``) when the target is known. Each input path is
-      classified by suffix into an ``object_file`` or ``static_library`` node
-      (best-effort textual classification, no archive introspection) and
-      connected via ``LINK_UNIT_HAS_INPUT`` — an object a compile unit already
-      emitted (same path) lands on the *same* node instead of a disconnected
-      duplicate, so a change traced to one object correlates across both
-      slices. A non-empty ``version_script`` gets its own node
-      (``LINK_UNIT_USES_VERSION_SCRIPT``).
-    - ``linker_script``/``export_map``/``comdat_group`` stay reserved
-      (schema-only) — no normalized data source for those three yet.
-      ``archive_member``/``ARCHIVE_CONTAINS_OBJECT``/``OBJECT_DEFINES_SYMBOL``
-      are *not* populated here — that needs a real archive introspection
-      pass (:mod:`~abicheck.buildsource.archive_graph`, G29 Phase 5 item 6),
-      run separately over the ``static_library`` nodes this function
-      creates; this function only classifies a link input by filename
-      suffix, it never opens the archive.
-
-    ``LINK_UNIT_EXPORTS_SYMBOL`` (a link unit's own exported symbols) is added
-    by :func:`_augment_with_source_abi` instead, once ``BINARY_EXPORTS_SYMBOL``
-    resolves which symbols the owning target actually exports — this function
-    runs first (build-evidence-only, no ``source_abi`` required) so the
-    ``link_unit`` node it creates is already there for that later step to
-    attach to.
-    """
-    for cu in build.compile_units:
-        if not cu.output:
-            continue
-        oid = _object_node_id(cu.output)
-        if not graph.has_node(oid):
-            graph.add_node(
-                GraphNode(
-                    id=oid,
-                    kind="object_file",
-                    label=cu.output,
-                    provenance="build_evidence",
-                    confidence=CONF_HIGH,
-                )
-            )
-        graph.add_edge(
-            GraphEdge(
-                src=cu.id,
-                dst=oid,
-                kind="COMPILE_UNIT_EMITS_OBJECT",
-                provenance="build_evidence",
-                confidence=CONF_HIGH,
-            )
-        )
-
-    known_targets = {t.id for t in build.targets}
-    for link in build.link_units:
-        graph.add_node(
-            GraphNode(
-                id=link.id,
-                kind="link_unit",
-                label=link.output or link.id,
-                provenance="build_evidence",
-                confidence=CONF_HIGH,
-                attrs={
-                    "kind": link.kind,
-                    "target_id": link.target_id,
-                    "soname": link.soname,
-                },
-            )
-        )
-        if link.target_id and link.target_id in known_targets:
-            graph.add_edge(
-                GraphEdge(
-                    src=link.target_id,
-                    dst=link.id,
-                    kind="TARGET_HAS_LINK_UNIT",
-                    provenance="build_evidence",
-                    confidence=CONF_HIGH,
-                )
-            )
-        for inp in link.inputs:
-            if not inp:
-                continue
-            is_archive = inp.lower().endswith(_STATIC_LIBRARY_SUFFIXES)
-            iid = _static_library_node_id(inp) if is_archive else _object_node_id(inp)
-            if not graph.has_node(iid):
-                graph.add_node(
-                    GraphNode(
-                        id=iid,
-                        kind="static_library" if is_archive else "object_file",
-                        label=inp,
-                        provenance="build_evidence",
-                        confidence=CONF_REDUCED,
-                    )
-                )
-            graph.add_edge(
-                GraphEdge(
-                    src=link.id,
-                    dst=iid,
-                    kind="LINK_UNIT_HAS_INPUT",
-                    provenance="build_evidence",
-                    confidence=CONF_HIGH,
-                )
-            )
-        if link.version_script:
-            vid = _version_script_node_id(link.version_script)
-            graph.add_node(
-                GraphNode(
-                    id=vid,
-                    kind="version_script",
-                    label=link.version_script,
-                    provenance="build_evidence",
-                    confidence=CONF_HIGH,
-                )
-            )
-            graph.add_edge(
-                GraphEdge(
-                    src=link.id,
-                    dst=vid,
-                    kind="LINK_UNIT_USES_VERSION_SCRIPT",
-                    provenance="build_evidence",
-                    confidence=CONF_HIGH,
-                )
-            )
-
-
-# ── Phases 3-4: enrich the graph from the ADR-030 L4 source surface ─────────
-
-
-def _augment_with_source_abi(
-    graph: SourceGraphSummary,
-    surface: SourceAbiSurface,
-    project_files: frozenset[str] | None = None,
-) -> None:
-    """Fold a linked L4 source surface into *graph* (Phases 3-4).
-
-    Adds the public-reachability slice (declarations/types/macros, each linked
-    to the public header that declares it) and the source↔binary slice (decl →
-    exported symbol, type → debug type, target → exported symbol). All edges are
-    tagged ``provenance="source_abi"`` so a reachability claim always discloses
-    that it rests on source-replay evidence, not a binary diff (ADR-031 D9).
-
-    *project_files* (``project_source_files(build)``) is threaded
-    through to :func:`fold_source_edges` so a ``source_edges`` endpoint can be
-    marked ``defined_in_project`` the same way ``augment_graph_with_calls``/
-    ``augment_graph_with_types`` already do for the standalone replay passes.
-    """
-    target_id = surface.target_id
-    if target_id and not graph.has_node(target_id):
-        # The surface may name a target the build evidence did not enumerate
-        # (e.g. binary+headers-only collection). Materialize it so its symbols
-        # have an owner in the graph.
-        graph.add_node(
-            GraphNode(
-                id=target_id,
-                kind="target",
-                label=target_id,
-                provenance="source_abi",
-                confidence=CONF_REDUCED,
-            )
-        )
-
-    decl_to_sym: dict[str, str] = surface.mappings.get(
-        "source_decl_to_binary_symbol", {}
-    )
-    type_to_dbg: dict[str, str] = surface.mappings.get("source_type_to_debug_type", {})
-    # ADR-041 P1 #2: the link unit(s) _fold_link_provenance already created for
-    # this target (build-evidence-only, before this function ran) — so an
-    # exported symbol can also be attributed to the specific link step that
-    # produced it, not only the target as a whole.
-    link_unit_ids = [
-        n.id
-        for n in graph.nodes
-        if n.kind == "link_unit" and target_id and n.attrs.get("target_id") == target_id
-    ]
-
-    def export_symbol(symbol: str, confidence: str) -> str:
-        sid = _symbol_node_id(symbol)
-        graph.add_node(
-            GraphNode(
-                id=sid,
-                kind="binary_symbol",
-                label=symbol,
-                provenance="source_abi",
-                confidence=CONF_HIGH,
-            )
-        )
-        if target_id:
-            graph.add_edge(
-                GraphEdge(
-                    src=target_id,
-                    dst=sid,
-                    kind="BINARY_EXPORTS_SYMBOL",
-                    provenance="source_abi",
-                    confidence=confidence,
-                )
-            )
-        for link_id in link_unit_ids:
-            graph.add_edge(
-                GraphEdge(
-                    src=link_id,
-                    dst=sid,
-                    kind="LINK_UNIT_EXPORTS_SYMBOL",
-                    provenance="source_abi",
-                    confidence=confidence,
-                )
-            )
-        return sid
-
-    def header_declares(entity: SourceEntity, node_id: str, confidence: str) -> None:
-        loc = entity.source_location
-        if loc is None or not loc.path:
-            return
-        hid = _header_node_id(loc.path)
-        # add_node keeps the first writer's facts, so a build-evidence header
-        # node (HIGH confidence) is not downgraded by this source_abi one.
-        graph.add_node(
-            GraphNode(
-                id=hid,
-                kind="header",
-                label=loc.path,
-                provenance="source_abi",
-                confidence=confidence,
-                attrs={"origin": loc.origin},
-            )
-        )
-        graph.add_edge(
-            GraphEdge(
-                src=hid,
-                dst=node_id,
-                kind="SOURCE_DECLARES",
-                provenance="source_abi",
-                confidence=confidence,
-            )
-        )
-
-    # Represent every exported symbol the surface mapped, so the target's export
-    # set is visible even for symbols whose declaration was not reachable.
-    for symbol in decl_to_sym.values():
-        if symbol:
-            export_symbol(symbol, CONF_REDUCED)
-    # source_link.py accounts for a real export under several other mappings
-    # too (each keyed *by* the symbol, unlike decl_to_sym): template-
-    # instantiation/synthesized/allocator-interposer/undocumented exports.
-    # Omitting these left no binary_symbol node for a downstream pass's
-    # join-only-onto-an-existing-node rule to find (Codex review).
-    for mapping_name in (
-        "template_instantiation_symbol_to_decl",
-        "synthesized_symbol_to_owner",
-        "allocator_interposer_symbol_to_owner",
-        "non_public_symbol_to_reason",
-    ):
-        for symbol in surface.mappings.get(mapping_name, {}):
-            if symbol:
-                export_symbol(symbol, CONF_REDUCED)
-
-    declarations = (
-        *surface.reachable_declarations,
-        *surface.reachable_templates,
-        *surface.reachable_inline_bodies,
-    )
-    # An entity routed to reachable_templates/reachable_inline_bodies shares its identity()
-    # (mangled name, or qualified_name+signature_hash) with the plain "function"-kind
-    # declaration entity clang.py *also* emits for the same function -- both land on the same
-    # node id via _decl_node_id below, and add_node keeps only the first writer's attrs.
-    # reachable_declarations is iterated first in `declarations` above, so for any function
-    # that also has an inline/template rendition, the winning node's own
-    # attrs["decl_kind"] is always "function"/"method", never "inline"/"template" -- silently
-    # losing the one signal that distinguishes "body compiled into every consumer TU that
-    # includes this header" from "ordinary out-of-line body, compiled into this library's
-    # binary only" (Codex review). Compute the identity set up front so every entity sharing
-    # it gets the *same* attrs["consumer_compiled_body"] value regardless of which one wins.
-    consumer_compiled_identities = {
-        ent.identity()
-        for ent in (*surface.reachable_templates, *surface.reachable_inline_bodies)
-    }
-    for ent in declarations:
-        did = _decl_node_id(ent.identity())
-        conf = ent.confidence.value
-        graph.add_node(
-            GraphNode(
-                id=did,
-                kind="source_decl",
-                label=ent.qualified_name or ent.identity(),
-                provenance="source_abi",
-                confidence=conf,
-                attrs={
-                    "decl_kind": ent.kind,
-                    "visibility": ent.visibility,
-                    "consumer_compiled_body": ent.identity()
-                    in consumer_compiled_identities,
-                },
-            )
-        )
-        header_declares(ent, did, conf)
-        # decl_to_sym is keyed by entity identity (the mangled name for C++, so
-        # overloads stay distinct) by both link_source_abi and
-        # relink_surface_exports — look it up the same way, not by qualified_name,
-        # or the SOURCE_DECL_MAPS_TO_SYMBOL edge is never created for C++.
-        symbol = decl_to_sym.get(ent.identity(), "")
-        if symbol:
-            graph.add_edge(
-                GraphEdge(
-                    src=did,
-                    dst=_symbol_node_id(symbol),
-                    kind="SOURCE_DECL_MAPS_TO_SYMBOL",
-                    provenance="source_abi",
-                    confidence=conf,
-                )
-            )
-
-    for ent in surface.reachable_types:
-        tid = _type_node_id(ent.identity())
-        conf = ent.confidence.value
-        graph.add_node(
-            GraphNode(
-                id=tid,
-                kind=_type_node_kind(ent.kind),
-                label=ent.qualified_name or ent.identity(),
-                provenance="source_abi",
-                confidence=conf,
-                attrs={"decl_kind": ent.kind, "visibility": ent.visibility},
-            )
-        )
-        header_declares(ent, tid, conf)
-        debug_type = type_to_dbg.get(ent.qualified_name, "")
-        if debug_type:
-            bid = _debug_type_node_id(debug_type)
-            graph.add_node(
-                GraphNode(
-                    id=bid,
-                    kind="debug_type",
-                    label=debug_type,
-                    provenance="source_abi",
-                    confidence=CONF_REDUCED,
-                )
-            )
-            graph.add_edge(
-                GraphEdge(
-                    src=tid,
-                    dst=bid,
-                    kind="SOURCE_TYPE_MAPS_TO_DEBUG_TYPE",
-                    provenance="source_abi",
-                    confidence=CONF_REDUCED,
-                )
-            )
-
-    for ent in surface.reachable_macros:
-        mid = _macro_node_id(ent.qualified_name or ent.identity())
-        conf = ent.confidence.value
-        graph.add_node(
-            GraphNode(
-                id=mid,
-                kind="macro",
-                label=ent.qualified_name or ent.identity(),
-                provenance="source_abi",
-                confidence=conf,
-            )
-        )
-        header_declares(ent, mid, conf)
-
-    fold_source_edges(graph, surface.source_edges, project_files)
-
-
-def _source_edge_endpoint_ids(
-    kind: str, src: str, dst: str
-) -> tuple[str, str, str, str]:
-    """Map a raw ``source_edges`` row's ``(kind, src, dst)`` identities onto
-    graph node ids/kinds, mirroring the id scheme
-    ``call_graph.augment_graph_with_calls``/``type_graph.augment_graph_with_types``
-    already use — so an edge folded from L4 facts lands on the same
-    ``decl://``/``type://`` node a separate call/type-graph replay pass (or L4
-    declaration enrichment) would have created, rather than a disconnected
-    duplicate.
-    """
-    if kind == "DECL_HAS_TYPE":
-        return _decl_node_id(src), "source_decl", _type_node_id(dst), "record_type"
-    if kind in ("TYPE_INHERITS", "TYPE_HAS_FIELD_TYPE"):
-        return _type_node_id(src), "record_type", _type_node_id(dst), "record_type"
-    # DECL_CALLS_DECL / DECL_REFERENCES_DECL — the only other kinds a caller
-    # reaches this with (fold_source_edges gates on DEPENDENCY_EDGE_KINDS
-    # before calling this).
-    return _decl_node_id(src), "source_decl", _decl_node_id(dst), "source_decl"
-
-
-def fold_source_edges(
-    graph: SourceGraphSummary,
-    source_edges: list[dict[str, Any]],
-    project_files: frozenset[str] | None = None,
-) -> int:
-    """Fold ``SourceAbiSurface.source_edges`` into *graph* (ADR-038 C.9 / PR1).
-
-    Closes the gap where a Clang-plugin/replay-collected ``source_edges`` fact
-    was serialized onto ``SourceAbiTu``/``SourceAbiSurface`` but never reached
-    the L5 graph (latest-main Clang plugin review): ``DECL_CALLS_DECL``,
-    ``DECL_REFERENCES_DECL``, ``DECL_HAS_TYPE``, ``TYPE_HAS_FIELD_TYPE``, and
-    ``TYPE_INHERITS`` rows collected during the *same* L4 frontend invocation
-    are folded in exactly like a separate ``call_graph``/``type_graph`` replay
-    pass would, using the identical node-id scheme -- so an edge here
-    reconciles with (de-duplicates against, via ``add_edge``'s
-    ``(src, dst, kind)`` key) one already present from L4 declaration
-    enrichment or a separate replay pass, first-writer-wins.
-
-    Malformed rows (missing edge-kind/src/dst, or a non-dict entry from a
-    hand-edited/forward-versioned pack) are skipped rather than raising --
-    ``source_edges`` is best-effort collected evidence (ADR-028 D7), never a
-    reason to abort the rest of the graph build. Returns the number of edges
-    actually added (excludes rows that duplicated an edge already present).
-
-    When *project_files* is supplied and a row's ``attrs["dst_file"]`` matches
-    one of them, the dst node is marked ``defined_in_project`` (+ ``def_file``)
-    -- mirroring ``call_graph.augment_graph_with_calls``/
-    ``type_graph.augment_graph_with_types``'s identical marker for the
-    standalone replay passes. Without this, a callee/reference/type that only
-    ever appears as a ``source_edges`` endpoint (never independently declared
-    on the L4 public surface) carries no project provenance at all, so
-    ``is_internal_dependency_node`` cannot recognize it and
-    ``PUBLIC_API_INTERNAL_DEPENDENCY_ADDED`` silently misses it (Codex review
-    on PR #555; the exact gap this ADR's own "still always run [the replay]"
-    note flags as outstanding for the ``source_edges`` wire format). Which
-    rows carry ``dst_file`` depends on the producer: the Python inline
-    extractor (``clang_source_edges.py``) resolves it for every edge kind;
-    the ADR-038 C.8 clang plugin resolves it for all five kinds too as of
-    ADR-038 C.13 (a ``typeDeclFile(QualType)`` helper unwraps
-    pointer/reference/array sugar and resolves a typedef alias to its own
-    declaring file, or a record/enum ``TagDecl`` otherwise) -- though its
-    ``DECL_HAS_TYPE`` still never covers a variable's own type or a
-    typedef's underlying type (only function return/parameter types), so
-    ``mark_source_edges_extractor_coverage()`` still degrades the whole
-    family for the plugin producer rather than trusting it, per that
-    function's docstring. Applied whether the
-    node is created fresh here or already existed from an earlier edge in
-    this same call (backfilled, unless it already carries a ``visibility``
-    attr -- real L4 evidence, never overridden by this best-effort marker),
-    mirroring ``augment_graph_with_types``'s identical backfill behavior.
-    """
-    node_by_id: dict[str, GraphNode] = {n.id: n for n in graph.nodes}
-    added = 0
-    for row in source_edges:
-        if not isinstance(row, dict):
-            continue
-        kind = str(row.get("edge") or row.get("kind") or "")
-        src_ident = str(row.get("src") or "")
-        dst_ident = str(row.get("dst") or "")
-        # DEPENDENCY_EDGE_KINDS, not the broader EDGE_KINDS (CodeRabbit
-        # review, PR #555): source_edges only ever carries these five
-        # decl/type-dependency kinds, so a forward-incompatible or malformed
-        # row naming an unrelated kind (e.g. TARGET_DEPENDS_ON) must not
-        # silently fall through to the decl/decl default mapping below.
-        if not kind or not src_ident or not dst_ident:
-            continue
-        if kind not in DEPENDENCY_EDGE_KINDS:
-            continue
-        src_id, src_kind, dst_id, dst_kind = _source_edge_endpoint_ids(
-            kind, src_ident, dst_ident
-        )
-        confidence = str(row.get("confidence") or CONF_UNKNOWN)
-        provenance = str(row.get("provenance") or "source_edges")
-        attrs_raw = row.get("attrs")
-        row_attrs = dict(attrs_raw) if isinstance(attrs_raw, dict) else {}
-        dst_file = str(row_attrs.get("dst_file", ""))
-        dst_in_project = bool(
-            project_files and dst_file and _file_in_project(dst_file, project_files)
-        )
-        for node_id, node_kind, ident, is_dst in (
-            (src_id, src_kind, src_ident, False),
-            (dst_id, dst_kind, dst_ident, True),
-        ):
-            existing = node_by_id.get(node_id)
-            if existing is None:
-                node_attrs = (
-                    {"defined_in_project": True, "def_file": dst_file}
-                    if is_dst and dst_in_project
-                    else {}
-                )
-                node = GraphNode(
-                    id=node_id,
-                    kind=node_kind,
-                    label=ident,
-                    provenance=provenance,
-                    confidence=confidence,
-                    attrs=node_attrs,
-                )
-                graph.add_node(node)
-                node_by_id[node_id] = node
-            elif (
-                is_dst
-                and dst_in_project
-                and not existing.attrs.get("defined_in_project")
-                and not existing.attrs.get("visibility")
-            ):
-                # ADR-046 D2: route through register_fact (a direct
-                # existing.attrs[...] mutation is dropped on the next round-trip).
-                backfill = {"defined_in_project": True, "def_file": dst_file}
-                register_fact(existing, provenance, confidence, backfill)
-        before = len(graph.edges)
-        graph.add_edge(
-            GraphEdge(
-                src=src_id,
-                dst=dst_id,
-                kind=kind,
-                provenance=provenance,
-                confidence=confidence,
-                attrs=row_attrs,
-            )
-        )
-        if len(graph.edges) > before:
-            added += 1
-    return added
-
-
-def mark_source_edges_extractor_coverage(
-    graph: SourceGraphSummary, surface: SourceAbiSurface | None
-) -> None:
-    """Translate a confirmed-complete ``source_edges`` rollup into
-    ``call_graph``/``type_graph`` extractor-pass coverage (Codex review).
-
-    ``fold_source_edges`` (called from :func:`build_source_graph`) never
-    touches ``graph.extractor_passes`` itself -- when a caller runs the
-    ``call_graph``/``type_graph`` replay right after building the graph
-    (``inline._build_inline_graph``, ``cli_buildsource_helpers``), that
-    replay's own ``extractor_pass_fully_covered()``/``narrowed_pass_confirmed()``
-    tracking is strictly more precise (it knows full-vs-narrowed scope; a bare
-    ``source_edges`` rollup does not) and must be the sole source of truth —
-    do not call this alongside it. But a caller that folds ``source_edges``
-    and never runs a replay at all (``inputs_pack.ingest_inputs_pack``
-    ingesting a build-emitted Flow-2 pack; ``cli_buildsource_merge``'s
-    export-relink graph rebuild) leaves both flags permanently unset even
-    though the AST was genuinely, completely walked for these edge kinds --
-    ``source_graph_findings._common_dependency_edge_kinds``/
-    ``_dependency_kinds_covered`` then read that as "no pass ever ran" and
-    suppress a real ``PUBLIC_API_INTERNAL_DEPENDENCY_ADDED`` finding as a
-    coverage artifact instead of reporting it. A Flow-2 pack always reflects
-    whatever the build compiled (never a `changed_paths`-narrowed subset the
-    way an inline scan can be), so a confirmed-complete rollup here is safe
-    to treat as full-scope coverage.
-
-    "complete" is only trusted when ``surface.source_edges`` is actually
-    non-empty (Codex review, PR #555): ``coverage["fact_family_states"]``
-    predates ``SourceAbiSurface.source_edges`` (ADR-038 C.8 vs. C.9), so a
-    pre-C.9 persisted ``source_abi.json`` can carry ``source_edges:
-    "complete"`` from when the per-TU edges existed but its serializer had
-    no field to persist them into -- ``SourceAbiSurface.from_dict`` then
-    defaults the now-missing key to ``[]``. Treating that as confirmed-zero
-    coverage would read a schema-version gap as "nothing to see here",
-    letting a pre-existing internal dependency look newly added the moment
-    such a legacy baseline is compared against a freshly regenerated
-    candidate. A mismatched "complete"-with-no-edges is left unmarked here
-    (same as absent/unsupported), never silently upgraded.
-
-    Gated on the producer being ``_FULL_WALK_SOURCE_EDGES_PRODUCER`` (Codex
-    review, PR #555): "complete"/"empty-confirmed" only means "every TU's
-    ``source_edges`` collection ran without trouble", not "every function/type
-    in the TU was walked". The Python inline extractor
-    (``clang_source_edges.build_source_edges``) reuses ``call_graph.py``'s/
-    ``type_graph.py``'s full, unfiltered AST walk, so its coverage genuinely
-    matches a standalone replay. The ADR-038 C.8 clang plugin's ``source_edges``
-    does not: ``VisitFunctionDecl`` returns before running ``CallRefVisitor``
-    unless ``classify()`` accepts the function (public-header-declared only --
-    a private/internal helper defined purely in a ``.cpp`` is skipped
-    entirely, its outgoing calls never walked), and it never emits
-    ``DECL_HAS_TYPE`` for a typedef's underlying type or a variable's type (only
-    for function return/parameter types) at all. Aliasing the plugin's
-    ``source_edges`` to full ``call_graph``/``type_graph`` trust would read
-    "the public surface's calls/types were captured" as "the whole TU's
-    call/type graph is confirmed empty beyond what's here" -- hiding a
-    genuinely new dependency added inside a private helper's body, or a
-    changed typedef/variable type, as a false negative. A rolled-up
-    ``fact_set`` that disagrees across TUs, or is missing (pre-C.8 producer,
-    mixed pack), is treated the same as the plugin case: never grant blanket
-    trust without a positive, unambiguous "full walk" signal.
-
-    A non-full-walk producer (or an unresolved one) whose ``source_edges``
-    nonetheless folded real edges into *graph* is stamped ``degraded_passes``
-    instead of left entirely unmarked (Codex review): an unmarked pass falls
-    back to raw edge *presence* in
-    ``source_graph_findings._common_dependency_edge_kinds`` (its
-    ``_pass_ran``/``_pass_trusted_kinds`` checks only consult
-    ``extractor_passes``/``narrowed_passes``, not the *absence* of a
-    ``degraded_passes`` entry) — and a scoped producer's edges cannot safely
-    vouch for a project-wide zero any more than a narrowed/degraded
-    standalone replay's edges can (the same one-directional risk the sixth/
-    sixteenth Codex reviews already established ``degraded_passes`` guards
-    against elsewhere in that module). Left unmarked, a plugin baseline with
-    even one public-surface call edge would make ``DECL_CALLS_DECL`` look
-    "common" against a full-replay candidate, and a pre-existing
-    private-helper dependency the plugin structurally could never have seen
-    would surface as a false ``PUBLIC_API_INTERNAL_DEPENDENCY_ADDED`` the
-    moment collection switches producers. ``degraded_passes`` only ever
-    restricts trust in *this* side's absence of a kind (never gates the
-    *other* side's presence), so this can only trade a missed addition for
-    avoiding a false alarm — the same conservative bias the whole
-    narrowed/degraded chain already commits to.
-    """
-    if surface is None:
-        return
-    families = surface.coverage.get("fact_family_states")
-    # A missing/malformed fact_family_states (a third-party or hand-edited
-    # surface, or a schema older than ADR-038 C.8) must not fall through to
-    # "return unmarked" when source_edges nonetheless folded real edges into
-    # *graph* -- that leaves the exact same raw-edge-presence-fallback gap
-    # a non-full-walk producer does (Codex review): treated as unknown/
-    # non-full coverage below (state stays None, so the full-walk-trust
-    # branch never fires), falling through to the degraded stamp instead of
-    # returning early.
-    state = families.get("source_edges") if isinstance(families, dict) else None
-    fact_set = surface.coverage.get("fact_set")
-    full_walk_producer = (
-        isinstance(fact_set, dict)
-        and fact_set.get("producer") == _FULL_WALK_SOURCE_EDGES_PRODUCER
-    )
-    if full_walk_producer and (
-        state == "empty-confirmed" or (state == "complete" and surface.source_edges)
-    ):
-        graph.extractor_passes["call_graph"] = True
-        graph.extractor_passes["type_graph"] = True
-        return
-    if surface.source_edges:
-        graph.degraded_passes["call_graph"] = True
-        graph.degraded_passes["type_graph"] = True
-
-
-# ── Phase 5 (seed): structural graph-to-graph diff ──────────────────────────
-
-
-def _label_map(graph: SourceGraphSummary) -> dict[str, str]:
-    return {n.id: (n.label or n.id) for n in graph.nodes}
-
-
-def _kind_map(graph: SourceGraphSummary) -> dict[str, str]:
-    return {n.id: n.kind for n in graph.nodes}
-
-
-def localize_symbol(graph: SourceGraphSummary, symbol: str) -> dict[str, Any]:
-    """Localize an exported symbol through the graph (ADR-031 D8 `graph explain`).
-
-    Given a (mangled) binary symbol, walk the graph to report what produced and
-    reaches it: the exporting target(s), the source declaration(s) it maps to,
-    the public header(s) that declare those decls, the ABI-relevant build
-    option(s) that feed it, and the static callees of its declarations. Every
-    fact is graph-derived (provenance/confidence live on the edges), so the
-    result is explanatory, never an ABI verdict (ADR-031 D6).
-    """
-    labels = _label_map(graph)
-    kinds = _kind_map(graph)
-    sym_id = _symbol_node_id(symbol)
-    found = graph.has_node(sym_id)
-
-    targets = sorted(
-        {
-            e.src
-            for e in graph.edges
-            if e.kind == "BINARY_EXPORTS_SYMBOL" and e.dst == sym_id
-        }
-    )
-    decls = sorted(
-        {
-            e.src
-            for e in graph.edges
-            if e.kind == "SOURCE_DECL_MAPS_TO_SYMBOL" and e.dst == sym_id
-        }
-    )
-    options = sorted(
-        {
-            e.src
-            for e in graph.edges
-            if e.kind == "BUILD_OPTION_AFFECTS_SYMBOL" and e.dst == sym_id
-        }
-    )
-
-    headers: set[str] = set()
-    callees: set[str] = set()
-    for decl in decls:
-        headers |= {
-            e.src for e in graph.edges if e.kind == "SOURCE_DECLARES" and e.dst == decl
-        }
-        callees |= {
-            e.dst for e in graph.edges if e.kind == "DECL_CALLS_DECL" and e.src == decl
-        }
-
-    def names(ids: set[str] | list[str]) -> list[str]:
-        return sorted(labels.get(i, i) for i in ids)
-
-    return {
-        "symbol": symbol,
-        "found": found,
-        "exported_by_targets": names(targets),
-        "source_declarations": names(decls),
-        "declared_in_headers": names(headers),
-        "reached_by_build_options": names(options),
-        "static_callees": names(callees),
-        "header_kinds": {labels.get(h, h): kinds.get(h, "") for h in headers},
-    }
-
-
-def diff_source_graph(
-    old: SourceGraphSummary, new: SourceGraphSummary
-) -> GraphSummaryDiff:
-    """Compute the structural delta from *old* to *new* (Phase 5 seed).
-
-    Edge comparison deliberately stays keyed on the coarse ``key()``
-    (ADR-046 D1 — "existing callers... are unaffected"), not
-    ``relation_key()``: when two role-distinct edges share a ``(src, dst,
-    kind)`` (e.g. a function that both returns and takes the same private
-    type), only one is a "representative" for this structural added/removed
-    comparison — role-level diff granularity is not implemented here.
-    """
-    old_nodes = {n.id: n for n in old.nodes}
-    new_nodes = {n.id: n for n in new.nodes}
-    old_edges = {e.key(): e for e in old.edges}
-    new_edges = {e.key(): e for e in new.edges}
-
-    return GraphSummaryDiff(
-        added_nodes=[new_nodes[i] for i in sorted(new_nodes.keys() - old_nodes.keys())],
-        removed_nodes=[
-            old_nodes[i] for i in sorted(old_nodes.keys() - new_nodes.keys())
-        ],
-        added_edges=[new_edges[k] for k in sorted(new_edges.keys() - old_edges.keys())],
-        removed_edges=[
-            old_edges[k] for k in sorted(old_edges.keys() - new_edges.keys())
-        ],
-    )
+# Shared node/edge-classification predicates neither half owns exclusively.
+from .source_graph_query import (
+    DECL_NODE_KINDS as DECL_NODE_KINDS,
+    INTERNAL_VISIBILITIES as INTERNAL_VISIBILITIES,
+    PUBLIC_VISIBILITIES as PUBLIC_VISIBILITIES,
+    UNANNOTATED_VISIBILITIES as UNANNOTATED_VISIBILITIES,
+    decl_declaring_files as decl_declaring_files,
+    is_consumer_compiled_node as is_consumer_compiled_node,
+    is_consumer_compiled_public_entry as is_consumer_compiled_public_entry,
+    is_internal_dependency_node as is_internal_dependency_node,
+    is_public_dependency_node as is_public_dependency_node,
+    looks_like_system_name as looks_like_system_name,
+)
 
 
 # ── Back-compat re-export shim (lazy, to avoid an import cycle) ───────────────
