+++ abicheck/buildsource/source_graph_findings.py
@@ -35,19 +35,17 @@ from __future__ import annotations
 from collections import deque
 from typing import TYPE_CHECKING
 
+from ..model.graph_facts import GraphEdge
+from ..model.source_graph import EVIDENCE_TIER_L5, SourceGraphSummary
 from .header_graph import (
     HEADER_CALL_GRAPH_PASS,
     HEADER_INCLUDE_GRAPH_PASS,
     HEADER_TYPE_GRAPH_PASS,
 )
-from .source_graph import (
+from .source_graph_compare import _kind_map, _label_map
+from .source_graph_query import (
     _TYPE_ENTITY_KINDS,
-    EVIDENCE_TIER_L5,
     PUBLIC_VISIBILITIES,
-    GraphEdge,
-    SourceGraphSummary,
-    _kind_map,
-    _label_map,
     decl_declaring_files,
     is_internal_dependency_node,
     is_public_dependency_node,
