+++ architecture/debt.yaml
@@ -272,16 +272,6 @@
       "review_by": "2026-11-30"
     },
     {
-      "path": "abicheck/buildsource/source_graph.py",
-      "baseline_lines": 1352,
-      "target": "extract/build-or-source",
-      "rule": "no_growth",
-      "category": "legacy_extraction_monolith",
-      "owner": "abicheck maintainers",
-      "rationale": "ADR-061 Phase 5 item 2's \"values\" slice moved SourceGraphSummary/GraphSummaryDiff/node-id constructors/schema vocabulary to abicheck/model/source_graph.py (2000 -> 1352 lines); the remaining construction (build_source_graph) and comparison (diff_source_graph) halves predate ADR-061 and cannot move safely without their own behavior-preserving vertical slice.",
-      "review_by": "2026-11-30"
-    },
-    {
       "path": "abicheck/buildsource/source_graph_findings.py",
       "baseline_lines": 1812,
       "target": "extract/build-or-source",
@@ -2164,16 +2154,6 @@
     {
       "path": "tests/test_source_extractors_env.py",
       "baseline_lines": 1417,
-      "target": "tests/unit-or-contract ownership matching production migration",
-      "rule": "no_growth",
-      "category": "legacy_oversized_test_module",
-      "owner": "abicheck maintainers",
-      "rationale": "Existing test module predates ADR-061 and will split by behavior or contract axis with its production migration.",
-      "review_by": "2026-11-30"
-    },
-    {
-      "path": "tests/test_source_graph.py",
-      "baseline_lines": 1996,
       "target": "tests/unit-or-contract ownership matching production migration",
       "rule": "no_growth",
       "category": "legacy_oversized_test_module",
