+++ architecture/modules.yaml
@@ -85,6 +85,8 @@
         "abicheck/buildsource/redaction.py",
         "abicheck/buildsource/risk.py",
         "abicheck/buildsource/snapshot_exports.py",
+        "abicheck/buildsource/source_graph_build.py",
+        "abicheck/buildsource/source_graph_build_source_abi.py",
         "abicheck/buildsource/source_link.py",
         "abicheck/buildsource/source_replay.py",
         "abicheck/buildsource/template_graph_extractor.py",
@@ -160,6 +162,7 @@
         "abicheck/buildsource/crosscheck_coherence.py",
         "abicheck/buildsource/graph_reconcile.py",
         "abicheck/buildsource/source_diff.py",
+        "abicheck/buildsource/source_graph_compare.py",
         "abicheck/confidence.py",
         "abicheck/detector_registry.py",
         "abicheck/diff_abi_tags.py",
